+++ include/lwip-esp8266/lwipopts.h
@@ -38,6 +38,8 @@
 #ifndef __LWIPOPTS_H__
 #define __LWIPOPTS_H__
 
+#include <stdlib.h> // rand() / LWIP_RAND
+
 /*
    -----------------------------------------------
    ---------- Platform specific locking ----------
@@ -49,41 +51,31 @@
  * critical regions during buffer allocation, deallocation and memory
  * allocation and deallocation.
  */
-#ifndef SYS_LIGHTWEIGHT_PROT
 #define SYS_LIGHTWEIGHT_PROT            0
-#endif
 
 /** 
  * NO_SYS==1: Provides VERY minimal functionality. Otherwise,
  * use lwIP facilities.
  */
-#ifndef NO_SYS
 #define NO_SYS                          1
-#endif
 
 /**
  * NO_SYS_NO_TIMERS==1: Drop support for sys_timeout when NO_SYS==1
  * Mainly for compatibility to old versions.
  */
-#ifndef NO_SYS_NO_TIMERS
 #define NO_SYS_NO_TIMERS                0
-#endif
 
 /**
  * MEMCPY: override this if you have a faster implementation at hand than the
  * one included in your C library
  */
-#ifndef MEMCPY
 #define MEMCPY(dst,src,len)             os_memcpy(dst,src,len)
-#endif
 
 /**
  * SMEMCPY: override this with care! Some compilers (e.g. gcc) can inline a
  * call to memcpy() if the length is known at compile time and is small.
  */
-#ifndef SMEMCPY
 #define SMEMCPY(dst,src,len)            os_memcpy(dst,src,len)
-#endif
 
 /*
    ------------------------------------
@@ -95,44 +87,34 @@
  * instead of the lwip internal allocator. Can save code size if you
  * already use it.
  */
-#ifndef MEM_LIBC_MALLOC
 #define MEM_LIBC_MALLOC                 1
-#endif
 
 /**
 * MEMP_MEM_MALLOC==1: Use mem_malloc/mem_free instead of the lwip pool allocator.
 * Especially useful with MEM_LIBC_MALLOC but handle with care regarding execution
 * speed and usage from interrupts!
 */
-#ifndef MEMP_MEM_MALLOC
 #define MEMP_MEM_MALLOC                 1
-#endif
 
 /**
  * MEM_ALIGNMENT: should be set to the alignment of the CPU
  *    4 byte alignment -> #define MEM_ALIGNMENT 4
  *    2 byte alignment -> #define MEM_ALIGNMENT 2
  */
-#ifndef MEM_ALIGNMENT
 #define MEM_ALIGNMENT                   4
-#endif
 
 /**
  * MEM_SIZE: the size of the heap memory. If the application will send
  * a lot of data that needs to be copied, this should be set high.
  */
-#ifndef MEM_SIZE
 #define MEM_SIZE                        16000
-#endif
 
 /**
  * MEMP_SEPARATE_POOLS: if defined to 1, each pool is placed in its own array.
  * This can be used to individually change the location of each pool.
  * Default is one big array for all pools
  */
-#ifndef MEMP_SEPARATE_POOLS
 #define MEMP_SEPARATE_POOLS             1
-#endif
 
 /**
  * MEMP_OVERFLOW_CHECK: memp overflow protection reserves a configurable
@@ -143,17 +125,13 @@
  *    MEMP_OVERFLOW_CHECK >= 2 checks each element in every pool every time
  *      memp_malloc() or memp_free() is called (useful but slow!)
  */
-#ifndef MEMP_OVERFLOW_CHECK
 #define MEMP_OVERFLOW_CHECK             0
-#endif
 
 /**
  * MEMP_SANITY_CHECK==1: run a sanity check after each memp_free() to make
  * sure that there are no cycles in the linked lists.
  */
-#ifndef MEMP_SANITY_CHECK
 #define MEMP_SANITY_CHECK               1
-#endif
 
 /**
  * MEM_USE_POOLS==1: Use an alternative to malloc() by allocating from a set
@@ -161,17 +139,13 @@
  * the smallest pool that can provide the length needed is returned.
  * To use this, MEMP_USE_CUSTOM_POOLS also has to be enabled.
  */
-#ifndef MEM_USE_POOLS
 #define MEM_USE_POOLS                   0
-#endif
 
 /**
  * MEM_USE_POOLS_TRY_BIGGER_POOL==1: if one malloc-pool is empty, try the next
  * bigger pool - WARNING: THIS MIGHT WASTE MEMORY but it can make a system more
  * reliable. */
-#ifndef MEM_USE_POOLS_TRY_BIGGER_POOL
 #define MEM_USE_POOLS_TRY_BIGGER_POOL   0
-#endif
 
 /**
  * MEMP_USE_CUSTOM_POOLS==1: whether to include a user file lwippools.h
@@ -179,9 +153,7 @@
  * by lwIP. If you set this to 1, you must have lwippools.h in your 
  * inlude path somewhere. 
  */
-#ifndef MEMP_USE_CUSTOM_POOLS
 #define MEMP_USE_CUSTOM_POOLS           0
-#endif
 
 /**
  * Set this to 1 if you want to free PBUF_RAM pbufs (or call mem_free()) from
@@ -201,9 +173,7 @@
  * - pbuf_free_callback(p);
  * - mem_free_callback(m);
  */
-#ifndef LWIP_ALLOW_MEM_FREE_FROM_OTHER_CONTEXT
 #define LWIP_ALLOW_MEM_FREE_FROM_OTHER_CONTEXT 0
-#endif
 
 /*
    ------------------------------------------------
@@ -215,58 +185,44 @@
  * If the application sends a lot of data out of ROM (or other static memory),
  * this should be set high.
  */
-#ifndef MEMP_NUM_PBUF
 #define MEMP_NUM_PBUF                   10
-#endif
 
 /**
  * MEMP_NUM_RAW_PCB: Number of raw connection PCBs
  * (requires the LWIP_RAW option)
  */
-#ifndef MEMP_NUM_RAW_PCB
 #define MEMP_NUM_RAW_PCB                4
-#endif
 
 /**
  * MEMP_NUM_UDP_PCB: the number of UDP protocol control blocks. One
  * per active UDP "connection".
  * (requires the LWIP_UDP option)
  */
-#ifndef MEMP_NUM_UDP_PCB
 #define MEMP_NUM_UDP_PCB                4
-#endif
 
 /**
  * MEMP_NUM_TCP_PCB: the number of simulatenously active TCP connections.
  * (requires the LWIP_TCP option)
  */
-#ifndef MEMP_NUM_TCP_PCB
 #define MEMP_NUM_TCP_PCB                (*(volatile uint32*)0x600011FC)
-#endif
 
 /**
  * MEMP_NUM_TCP_PCB_LISTEN: the number of listening TCP connections.
  * (requires the LWIP_TCP option)
  */
-#ifndef MEMP_NUM_TCP_PCB_LISTEN
 #define MEMP_NUM_TCP_PCB_LISTEN         2
-#endif
 
 /**
  * MEMP_NUM_TCP_SEG: the number of simultaneously queued TCP segments.
  * (requires the LWIP_TCP option)
  */
-#ifndef MEMP_NUM_TCP_SEG
 #define MEMP_NUM_TCP_SEG                16
-#endif
 
 /**
  * MEMP_NUM_REASSDATA: the number of simultaneously IP packets queued for
  * reassembly (whole packets, not fragments!)
  */
-#ifndef MEMP_NUM_REASSDATA
 #define MEMP_NUM_REASSDATA              0
-#endif
 
 /**
  * MEMP_NUM_FRAG_PBUF: the number of IP fragments simultaneously sent
@@ -275,9 +231,7 @@
  * LWIP_NETIF_TX_SINGLE_PBUF==0 and only has to be > 1 with DMA-enabled MACs
  * where the packet is not yet sent when netif->output returns.
  */
-#ifndef MEMP_NUM_FRAG_PBUF
 #define MEMP_NUM_FRAG_PBUF              0
-#endif
 
 /**
  * MEMP_NUM_ARP_QUEUE: the number of simulateously queued outgoing
@@ -285,9 +239,7 @@
  * their destination address) to finish.
  * (requires the ARP_QUEUEING option)
  */
-#ifndef MEMP_NUM_ARP_QUEUE
 #define MEMP_NUM_ARP_QUEUE              10
-#endif
 
 /**
  * MEMP_NUM_IGMP_GROUP: The number of multicast groups whose network interfaces
@@ -295,115 +247,87 @@
  * per netif membership).
  * (requires the LWIP_IGMP option)
  */
-#ifndef MEMP_NUM_IGMP_GROUP
 #define MEMP_NUM_IGMP_GROUP             8
-#endif
 
 /**
  * MEMP_NUM_SYS_TIMEOUT: the number of simulateously active timeouts.
  * (requires NO_SYS==0)
  */
-#ifndef MEMP_NUM_SYS_TIMEOUT
 #define MEMP_NUM_SYS_TIMEOUT            6
-#endif
 
 /**
  * MEMP_NUM_NETBUF: the number of struct netbufs.
  * (only needed if you use the sequential API, like api_lib.c)
  */
-#ifndef MEMP_NUM_NETBUF
 #define MEMP_NUM_NETBUF                 0
-#endif
 
 /**
  * MEMP_NUM_NETCONN: the number of struct netconns.
  * (only needed if you use the sequential API, like api_lib.c)
  */
-#ifndef MEMP_NUM_NETCONN
 #define MEMP_NUM_NETCONN                0
-#endif
 
 /**
  * MEMP_NUM_TCPIP_MSG_API: the number of struct tcpip_msg, which are used
  * for callback/timeout API communication. 
  * (only needed if you use tcpip.c)
  */
-#ifndef MEMP_NUM_TCPIP_MSG_API
 #define MEMP_NUM_TCPIP_MSG_API          4
-#endif
 
 /**
  * MEMP_NUM_TCPIP_MSG_INPKT: the number of struct tcpip_msg, which are used
  * for incoming packets. 
  * (only needed if you use tcpip.c)
  */
-#ifndef MEMP_NUM_TCPIP_MSG_INPKT
 #define MEMP_NUM_TCPIP_MSG_INPKT        4
-#endif
 
 /**
  * MEMP_NUM_SNMP_NODE: the number of leafs in the SNMP tree.
  */
-#ifndef MEMP_NUM_SNMP_NODE
 #define MEMP_NUM_SNMP_NODE              0
-#endif
 
 /**
  * MEMP_NUM_SNMP_ROOTNODE: the number of branches in the SNMP tree.
  * Every branch has one leaf (MEMP_NUM_SNMP_NODE) at least!
  */
-#ifndef MEMP_NUM_SNMP_ROOTNODE
 #define MEMP_NUM_SNMP_ROOTNODE          0
-#endif
 
 /**
  * MEMP_NUM_SNMP_VARBIND: the number of concurrent requests (does not have to
  * be changed normally) - 2 of these are used per request (1 for input,
  * 1 for output)
  */
-#ifndef MEMP_NUM_SNMP_VARBIND
 #define MEMP_NUM_SNMP_VARBIND           0
-#endif
 
 /**
  * MEMP_NUM_SNMP_VALUE: the number of OID or values concurrently used
  * (does not have to be changed normally) - 3 of these are used per request
  * (1 for the value read and 2 for OIDs - input and output)
  */
-#ifndef MEMP_NUM_SNMP_VALUE
 #define MEMP_NUM_SNMP_VALUE             0
-#endif
 
 /**
  * MEMP_NUM_NETDB: the number of concurrently running lwip_addrinfo() calls
  * (before freeing the corresponding memory using lwip_freeaddrinfo()).
  */
-#ifndef MEMP_NUM_NETDB
 #define MEMP_NUM_NETDB                  0
-#endif
 
 /**
  * MEMP_NUM_LOCALHOSTLIST: the number of host entries in the local host list
  * if DNS_LOCAL_HOSTLIST_IS_DYNAMIC==1.
  */
-#ifndef MEMP_NUM_LOCALHOSTLIST
 #define MEMP_NUM_LOCALHOSTLIST          0
-#endif
 
 /**
  * MEMP_NUM_PPPOE_INTERFACES: the number of concurrently active PPPoE
  * interfaces (only used with PPPOE_SUPPORT==1)
  */
-#ifndef MEMP_NUM_PPPOE_INTERFACES
 #define MEMP_NUM_PPPOE_INTERFACES       0
-#endif
 
 /**
  * PBUF_POOL_SIZE: the number of buffers in the pbuf pool. 
  */
-#ifndef PBUF_POOL_SIZE
 #define PBUF_POOL_SIZE                  10
-#endif
 
 /*
    ---------------------------------
@@ -413,16 +337,12 @@
 /**
  * LWIP_ARP==1: Enable ARP functionality.
  */
-#ifndef LWIP_ARP
 #define LWIP_ARP                        1
-#endif
 
 /**
  * ARP_TABLE_SIZE: Number of active MAC-IP address pairs cached.
  */
-#ifndef ARP_TABLE_SIZE
 #define ARP_TABLE_SIZE                  10
-#endif
 
 /**
  * ARP_QUEUEING==1: Multiple outgoing packets are queued during hardware address
@@ -431,9 +351,7 @@
  * startup time. Set this to 1 if you know your application sends more than one
  * packet in a row to an IP address that is not in the ARP cache.
  */
-#ifndef ARP_QUEUEING
 #define ARP_QUEUEING                    1
-#endif
 
 /**
  * ETHARP_TRUST_IP_MAC==1: Incoming IP packets cause the ARP table to be
@@ -445,9 +363,7 @@
  * The peer *is* in the ARP table if it requested our address before.
  * Also notice that this slows down input processing of every IP packet!
  */
-#ifndef ETHARP_TRUST_IP_MAC
 #define ETHARP_TRUST_IP_MAC             1
-#endif
 
 /**
  * ETHARP_SUPPORT_VLAN==1: support receiving ethernet packets with VLAN header.
@@ -455,32 +371,24 @@
  * If ETHARP_VLAN_CHECK is defined, only VLAN-traffic for this VLAN is accepted.
  * If ETHARP_VLAN_CHECK is not defined, all traffic is accepted.
  */
-#ifndef ETHARP_SUPPORT_VLAN
 #define ETHARP_SUPPORT_VLAN             0
-#endif
 
 /** LWIP_ETHERNET==1: enable ethernet support for PPPoE even though ARP
  * might be disabled
  */
-#ifndef LWIP_ETHERNET
 #define LWIP_ETHERNET                   (LWIP_ARP || PPPOE_SUPPORT)
-#endif
 
 /** ETH_PAD_SIZE: number of bytes added before the ethernet header to ensure
  * alignment of payload after that header. Since the header is 14 bytes long,
  * without this padding e.g. addresses in the IP header will not be aligned
  * on a 32-bit boundary, so setting this to 2 can speed up 32-bit-platforms.
  */
-#ifndef ETH_PAD_SIZE
 #define ETH_PAD_SIZE                    0
-#endif
 
 /** ETHARP_SUPPORT_STATIC_ENTRIES==1: enable code to support static ARP table
  * entries (using etharp_add_static_entry/etharp_remove_static_entry).
  */
-#ifndef ETHARP_SUPPORT_STATIC_ENTRIES
 #define ETHARP_SUPPORT_STATIC_ENTRIES   0
-#endif
 
 
 /*
@@ -493,45 +401,35 @@
  * interfaces. If you are going to run lwIP on a device with only one network
  * interface, define this to 0.
  */
-#ifndef IP_FORWARD
 #define IP_FORWARD                      0
-#endif
 
 /**
  * IP_OPTIONS_ALLOWED: Defines the behavior for IP options.
  *      IP_OPTIONS_ALLOWED==0: All packets with IP options are dropped.
  *      IP_OPTIONS_ALLOWED==1: IP options are allowed (but not parsed).
  */
-#ifndef IP_OPTIONS_ALLOWED
 #define IP_OPTIONS_ALLOWED              1
-#endif
 
 /**
  * IP_REASSEMBLY==1: Reassemble incoming fragmented IP packets. Note that
  * this option does not affect outgoing packet sizes, which can be controlled
  * via IP_FRAG.
  */
-#ifndef IP_REASSEMBLY
 #define IP_REASSEMBLY                   0
-#endif
 
 /**
  * IP_FRAG==1: Fragment outgoing IP packets if their size exceeds MTU. Note
  * that this option does not affect incoming packet sizes, which can be
  * controlled via IP_REASSEMBLY.
  */
-#ifndef IP_FRAG
 #define IP_FRAG                         0
-#endif
 
 /**
  * IP_REASS_MAXAGE: Maximum time (in multiples of IP_TMR_INTERVAL - so seconds, normally)
  * a fragmented IP packet waits for all fragments to arrive. If not all fragments arrived
  * in this time, the whole packet is discarded.
  */
-#ifndef IP_REASS_MAXAGE
 #define IP_REASS_MAXAGE                 0
-#endif
 
 /**
  * IP_REASS_MAX_PBUFS: Total maximum amount of pbufs waiting to be reassembled.
@@ -539,9 +437,7 @@
  * PBUF_POOL_SIZE > IP_REASS_MAX_PBUFS so that the stack is still able to receive
  * packets even if the maximum amount of fragments is enqueued for reassembly!
  */
-#ifndef IP_REASS_MAX_PBUFS
 #define IP_REASS_MAX_PBUFS              0
-#endif
 
 /**
  * IP_FRAG_USES_STATIC_BUF==1: Use a static MTU-sized buffer for IP
@@ -550,9 +446,7 @@
  * new PBUF_RAM pbufs are used for fragments).
  * ATTENTION: IP_FRAG_USES_STATIC_BUF==1 may not be used for DMA-enabled MACs!
  */
-#ifndef IP_FRAG_USES_STATIC_BUF
 #define IP_FRAG_USES_STATIC_BUF         1
-#endif
 
 /**
  * IP_FRAG_MAX_MTU: Assumed max MTU on any interface for IP frag buffer
@@ -565,26 +459,20 @@
 /**
  * IP_DEFAULT_TTL: Default value for Time-To-Live used by transport layers.
  */
-#ifndef IP_DEFAULT_TTL
 #define IP_DEFAULT_TTL                  255
-#endif
 
 /**
  * IP_SOF_BROADCAST=1: Use the SOF_BROADCAST field to enable broadcast
  * filter per pcb on udp and raw send operations. To enable broadcast filter
  * on recv operations, you also have to set IP_SOF_BROADCAST_RECV=1.
  */
-#ifndef IP_SOF_BROADCAST
 #define IP_SOF_BROADCAST                0
-#endif
 
 /**
  * IP_SOF_BROADCAST_RECV (requires IP_SOF_BROADCAST=1) enable the broadcast
  * filter on recv operations.
  */
-#ifndef IP_SOF_BROADCAST_RECV
 #define IP_SOF_BROADCAST_RECV           0
-#endif
 
 /*
    ----------------------------------
@@ -595,30 +483,22 @@
  * LWIP_ICMP==1: Enable ICMP module inside the IP stack.
  * Be careful, disable that make your product non-compliant to RFC1122
  */
-#ifndef LWIP_ICMP
 #define LWIP_ICMP                       1
-#endif
 
 /**
  * ICMP_TTL: Default value for Time-To-Live used by ICMP packets.
  */
-#ifndef ICMP_TTL
 #define ICMP_TTL                       (IP_DEFAULT_TTL)
-#endif
 
 /**
  * LWIP_BROADCAST_PING==1: respond to broadcast pings (default is unicast only)
  */
-#ifndef LWIP_BROADCAST_PING
 #define LWIP_BROADCAST_PING             0
-#endif
 
 /**
  * LWIP_MULTICAST_PING==1: respond to multicast pings (default is unicast only)
  */
-#ifndef LWIP_MULTICAST_PING
 #define LWIP_MULTICAST_PING             0
-#endif
 
 /*
    ---------------------------------
@@ -628,16 +508,12 @@
 /**
  * LWIP_RAW==1: Enable application layer to hook into the IP layer itself.
  */
-#ifndef LWIP_RAW
 #define LWIP_RAW                        0
-#endif
 
 /**
  * LWIP_RAW==1: Enable application layer to hook into the IP layer itself.
  */
-#ifndef RAW_TTL
 #define RAW_TTL                        (IP_DEFAULT_TTL)
-#endif
 
 /*
    ----------------------------------
@@ -647,16 +523,12 @@
 /**
  * LWIP_DHCP==1: Enable DHCP module.
  */
-#ifndef LWIP_DHCP
 #define LWIP_DHCP                       1
-#endif
 
 /**
  * DHCP_DOES_ARP_CHECK==1: Do an ARP check on the offered address.
  */
-#ifndef DHCP_DOES_ARP_CHECK
 #define DHCP_DOES_ARP_CHECK             ((LWIP_DHCP) && (LWIP_ARP))
-#endif
 
 /*
    ------------------------------------
@@ -666,17 +538,13 @@
 /**
  * LWIP_AUTOIP==1: Enable AUTOIP module.
  */
-#ifndef LWIP_AUTOIP
 #define LWIP_AUTOIP                     0
-#endif
 
 /**
  * LWIP_DHCP_AUTOIP_COOP==1: Allow DHCP and AUTOIP to be both enabled on
  * the same interface at the same time.
  */
-#ifndef LWIP_DHCP_AUTOIP_COOP
 #define LWIP_DHCP_AUTOIP_COOP           0
-#endif
 
 /**
  * LWIP_DHCP_AUTOIP_COOP_TRIES: Set to the number of DHCP DISCOVER probes
@@ -685,9 +553,7 @@
  * be prepared to handle a changing IP address when DHCP overrides
  * AutoIP.
  */
-#ifndef LWIP_DHCP_AUTOIP_COOP_TRIES
 #define LWIP_DHCP_AUTOIP_COOP_TRIES     9
-#endif
 
 /*
    ----------------------------------
@@ -698,66 +564,50 @@
  * LWIP_SNMP==1: Turn on SNMP module. UDP must be available for SNMP
  * transport.
  */
-#ifndef LWIP_SNMP
 #define LWIP_SNMP                       0
-#endif
 
 /**
  * SNMP_CONCURRENT_REQUESTS: Number of concurrent requests the module will
  * allow. At least one request buffer is required. 
  */
-#ifndef SNMP_CONCURRENT_REQUESTS
 #define SNMP_CONCURRENT_REQUESTS        0
-#endif
 
 /**
  * SNMP_TRAP_DESTINATIONS: Number of trap destinations. At least one trap
  * destination is required
  */
-#ifndef SNMP_TRAP_DESTINATIONS
 #define SNMP_TRAP_DESTINATIONS          0
-#endif
 
 /**
  * SNMP_PRIVATE_MIB: 
  */
-#ifndef SNMP_PRIVATE_MIB
 #define SNMP_PRIVATE_MIB                0
-#endif
 
 /**
  * Only allow SNMP write actions that are 'safe' (e.g. disabeling netifs is not
  * a safe action and disabled when SNMP_SAFE_REQUESTS = 1).
  * Unsafe requests are disabled by default!
  */
-#ifndef SNMP_SAFE_REQUESTS
 #define SNMP_SAFE_REQUESTS              0
-#endif
 
 /**
  * The maximum length of strings used. This affects the size of
  * MEMP_SNMP_VALUE elements.
  */
-#ifndef SNMP_MAX_OCTET_STRING_LEN
 #define SNMP_MAX_OCTET_STRING_LEN       127
-#endif
 
 /**
  * The maximum depth of the SNMP tree.
  * With private MIBs enabled, this depends on your MIB!
  * This affects the size of MEMP_SNMP_VALUE elements.
  */
-#ifndef SNMP_MAX_TREE_DEPTH
 #define SNMP_MAX_TREE_DEPTH             15
-#endif
 
 /**
  * The size of the MEMP_SNMP_VALUE elements, normally calculated from
  * SNMP_MAX_OCTET_STRING_LEN and SNMP_MAX_TREE_DEPTH.
  */
-#ifndef SNMP_MAX_VALUE_SIZE
 #define SNMP_MAX_VALUE_SIZE             LWIP_MAX((SNMP_MAX_OCTET_STRING_LEN)+1, sizeof(s32_t)*(SNMP_MAX_TREE_DEPTH))
-#endif
 
 /*
    ----------------------------------
@@ -767,9 +617,7 @@
 /**
  * LWIP_IGMP==1: Turn on IGMP module. 
  */
-#ifndef LWIP_IGMP
 #define LWIP_IGMP                       1
-#endif
 
 /*
    ----------------------------------
@@ -780,34 +628,22 @@
  * LWIP_DNS==1: Turn on DNS module. UDP must be available for DNS
  * transport.
  */
-#ifndef LWIP_DNS
 #define LWIP_DNS                        1
-#endif
 
 /** DNS maximum number of entries to maintain locally. */
-#ifndef DNS_TABLE_SIZE
 #define DNS_TABLE_SIZE                  4
-#endif
 
 /** DNS maximum host name length supported in the name table. */
-#ifndef DNS_MAX_NAME_LENGTH
 #define DNS_MAX_NAME_LENGTH             256
-#endif
 
 /** The maximum of DNS servers */
-#ifndef DNS_MAX_SERVERS
 #define DNS_MAX_SERVERS                 2
-#endif
 
 /** DNS do a name checking between the query and the response. */
-#ifndef DNS_DOES_NAME_CHECK
 #define DNS_DOES_NAME_CHECK             1
-#endif
 
 /** DNS message max. size. Default value is RFC compliant. */
-#ifndef DNS_MSG_SIZE
 #define DNS_MSG_SIZE                    512
-#endif
 
 /** DNS_LOCAL_HOSTLIST: Implements a local host-to-address list. If enabled,
  *  you have to define
@@ -819,15 +655,11 @@
  *  #define DNS_LOOKUP_LOCAL_EXTERN(x) extern u32_t my_lookup_function(const char *name)
  *  that returns the IP address or INADDR_NONE if not found.
  */
-#ifndef DNS_LOCAL_HOSTLIST
 #define DNS_LOCAL_HOSTLIST              0
-#endif /* DNS_LOCAL_HOSTLIST */
 
 /** If this is turned on, the local host-list can be dynamically changed
  *  at runtime. */
-#ifndef DNS_LOCAL_HOSTLIST_IS_DYNAMIC
 #define DNS_LOCAL_HOSTLIST_IS_DYNAMIC   0
-#endif /* DNS_LOCAL_HOSTLIST_IS_DYNAMIC */
 
 /*
    ---------------------------------
@@ -837,30 +669,22 @@
 /**
  * LWIP_UDP==1: Turn on UDP.
  */
-#ifndef LWIP_UDP
 #define LWIP_UDP                        1
-#endif
 
 /**
  * LWIP_UDPLITE==1: Turn on UDP-Lite. (Requires LWIP_UDP)
  */
-#ifndef LWIP_UDPLITE
 #define LWIP_UDPLITE                    0
-#endif
 
 /**
  * UDP_TTL: Default Time-To-Live value.
  */
-#ifndef UDP_TTL
 #define UDP_TTL                         (IP_DEFAULT_TTL)
-#endif
 
 /**
  * LWIP_NETBUF_RECVINFO==1: append destination addr and port to every netbuf.
  */
-#ifndef LWIP_NETBUF_RECVINFO
 #define LWIP_NETBUF_RECVINFO            0
-#endif
 
 /*
    ---------------------------------
@@ -870,61 +694,45 @@
 /**
  * LWIP_TCP==1: Turn on TCP.
  */
-#ifndef LWIP_TCP
 #define LWIP_TCP                        1
-#endif
 
 /**
  * TCP_TTL: Default Time-To-Live value.
  */
-#ifndef TCP_TTL
 #define TCP_TTL                         (IP_DEFAULT_TTL)
-#endif
 
 /**
  * TCP_WND: The size of a TCP window.  This must be at least 
  * (2 * TCP_MSS) for things to work well
  */
-//#ifndef TCP_WND
 //#define TCP_WND                         (4 * TCP_MSS)
 #define TCP_WND                         (2 * TCP_MSS)
-//#endif 
 
 /**
  * TCP_MAXRTX: Maximum number of retransmissions of data segments.
  */
-#ifndef TCP_MAXRTX
 #define TCP_MAXRTX                      3
-#endif
 
 /**
  * TCP_SYNMAXRTX: Maximum number of retransmissions of SYN segments.
  */
-#ifndef TCP_SYNMAXRTX
 #define TCP_SYNMAXRTX                   3
-#endif
 
 /**
  * TCP_MAXRTO: Maximum retransmission timeout of data segments.
  */
-#ifndef TCP_MAXRTO
 #define TCP_MAXRTO                      10
-#endif
 
 /**
  * TCP_MINRTO: Minimum retransmission timeout of data segments.
  */
-#ifndef TCP_MINRTO
 #define TCP_MINRTO                      2
-#endif
 
 /**
  * TCP_QUEUE_OOSEQ==1: TCP will queue segments that arrive out of order.
  * Define to 0 if your device is low on memory.
  */
-#ifndef TCP_QUEUE_OOSEQ
 #define TCP_QUEUE_OOSEQ                 0
-#endif
 
 /**
  * TCP_MSS: TCP Maximum segment size. (default is 536, a conservative default,
@@ -933,11 +741,9 @@
  * when opening a connection. For the transmit size, this MSS sets
  * an upper limit on the MSS advertised by the remote host.
  */
-//#ifndef TCP_MSS
 //#define TCP_MSS                         1460
 //#define TCP_MSS                         536
 #define TCP_MSS                         1024
-//#endif
 
 /**
  * TCP_CALCULATE_EFF_SEND_MSS: "The maximum size of a segment that TCP really
@@ -947,59 +753,45 @@
  * Setting this to 1 enables code that checks TCP_MSS against the MTU of the
  * netif used for a connection and limits the MSS if it would be too big otherwise.
  */
-#ifndef TCP_CALCULATE_EFF_SEND_MSS
 #define TCP_CALCULATE_EFF_SEND_MSS      1
-#endif
 
 
 /**
  * TCP_SND_BUF: TCP sender buffer space (bytes). 
  */
-#ifndef TCP_SND_BUF
 #define TCP_SND_BUF                     2 * TCP_MSS
-#endif
 
 /**
  * TCP_SND_QUEUELEN: TCP sender buffer space (pbufs). This must be at least
  * as much as (2 * TCP_SND_BUF/TCP_MSS) for things to work.
  */
-#ifndef TCP_SND_QUEUELEN
 #define TCP_SND_QUEUELEN                ((4 * (TCP_SND_BUF) + (TCP_MSS - 1))/(TCP_MSS))
-#endif
 
 /**
  * TCP_SNDLOWAT: TCP writable space (bytes). This must be less than
  * TCP_SND_BUF. It is the amount of space which must be available in the
  * TCP snd_buf for select to return writable (combined with TCP_SNDQUEUELOWAT).
  */
-#ifndef TCP_SNDLOWAT
 #define TCP_SNDLOWAT                    ((TCP_SND_BUF)/2)
-#endif
 
 /**
  * TCP_SNDQUEUELOWAT: TCP writable bufs (pbuf count). This must be grater
  * than TCP_SND_QUEUELEN. If the number of pbufs queued on a pcb drops below
  * this number, select returns writable (combined with TCP_SNDLOWAT).
  */
-#ifndef TCP_SNDQUEUELOWAT
 #define TCP_SNDQUEUELOWAT               LWIP_MAX(((TCP_SND_QUEUELEN)/2), 5)
-#endif
 
 /**
  * TCP_LISTEN_BACKLOG: Enable the backlog option for tcp listen pcb.
  */
-#ifndef TCP_LISTEN_BACKLOG
 #define TCP_LISTEN_BACKLOG              0
-#endif
 
 /**
  * The maximum allowed backlog for TCP listen netconns.
  * This backlog is used unless another is explicitly specified.
  * 0xff is the maximum (u8_t).
  */
-#ifndef TCP_DEFAULT_LISTEN_BACKLOG
 #define TCP_DEFAULT_LISTEN_BACKLOG      0xff
-#endif
 
 /**
  * TCP_OVERSIZE: The maximum number of bytes that tcp_write may
@@ -1015,24 +807,18 @@
  * TCP_MSS:   Try to create unfragmented TCP packets.
  * TCP_MSS/4: Try to create 4 fragments or less per TCP packet.
  */
-#ifndef TCP_OVERSIZE
 #define TCP_OVERSIZE                    TCP_MSS
-#endif
 
 /**
  * LWIP_TCP_TIMESTAMPS==1: support the TCP timestamp option.
  */
-#ifndef LWIP_TCP_TIMESTAMPS
 #define LWIP_TCP_TIMESTAMPS             0
-#endif
 
 /**
  * TCP_WND_UPDATE_THRESHOLD: difference in window to trigger an
  * explicit window update
  */
-#ifndef TCP_WND_UPDATE_THRESHOLD
 #define TCP_WND_UPDATE_THRESHOLD   (TCP_WND / 4)
-#endif
 
 /**
  * LWIP_EVENT_API and LWIP_CALLBACK_API: Only one of these should be set to 1.
@@ -1041,13 +827,8 @@
  *     LWIP_CALLBACK_API==1: The PCB callback function is called directly
  *         for the event.
  */
-#ifndef LWIP_EVENT_API
 #define LWIP_EVENT_API                  0
 #define LWIP_CALLBACK_API               1
-#else 
-#define LWIP_EVENT_API                  1
-#define LWIP_CALLBACK_API               0
-#endif
 
 
 /*
@@ -1060,18 +841,14 @@
  * link level header. The default is 14, the standard value for
  * Ethernet.
  */
-#ifndef PBUF_LINK_HLEN
 #define PBUF_LINK_HLEN                  (14 + ETH_PAD_SIZE)
-#endif
 
 /**
  * PBUF_POOL_BUFSIZE: the size of each pbuf in the pbuf pool. The default is
  * designed to accomodate single full size TCP frame in one pbuf, including
  * TCP_MSS, IP header, and link header.
  */
-#ifndef PBUF_POOL_BUFSIZE
 #define PBUF_POOL_BUFSIZE               LWIP_MEM_ALIGN_SIZE(TCP_MSS+40+PBUF_LINK_HLEN)
-#endif
 
 /*
    ------------------------------------------------
@@ -1082,32 +859,24 @@
  * LWIP_NETIF_HOSTNAME==1: use DHCP_OPTION_HOSTNAME with netif's hostname
  * field.
  */
-#ifndef LWIP_NETIF_HOSTNAME
 #define LWIP_NETIF_HOSTNAME             0
-#endif
 
 /**
  * LWIP_NETIF_API==1: Support netif api (in netifapi.c)
  */
-#ifndef LWIP_NETIF_API
 #define LWIP_NETIF_API                  0
-#endif
 
 /**
  * LWIP_NETIF_STATUS_CALLBACK==1: Support a callback function whenever an interface
  * changes its up/down status (i.e., due to DHCP IP acquistion)
  */
-#ifndef LWIP_NETIF_STATUS_CALLBACK
 #define LWIP_NETIF_STATUS_CALLBACK      0
-#endif
 
 /**
  * LWIP_NETIF_LINK_CALLBACK==1: Support a callback function from an interface
  * whenever the link changes (i.e., link down)
  */
-#ifndef LWIP_NETIF_LINK_CALLBACK
 #define LWIP_NETIF_LINK_CALLBACK        0
-#endif
 
 /**
  * LWIP_NETIF_HWADDRHINT==1: Cache link-layer-address hints (e.g. table
@@ -1116,25 +885,19 @@
  * ARP tables or many concurrent connections, it might be counterproductive
  * if you have a tiny ARP table or if there never are concurrent connections.
  */
-#ifndef LWIP_NETIF_HWADDRHINT
 #define LWIP_NETIF_HWADDRHINT           0
-#endif
 
 /**
  * LWIP_NETIF_LOOPBACK==1: Support sending packets with a destination IP
  * address equal to the netif IP address, looping them back up the stack.
  */
-#ifndef LWIP_NETIF_LOOPBACK
 #define LWIP_NETIF_LOOPBACK             0
-#endif
 
 /**
  * LWIP_LOOPBACK_MAX_PBUFS: Maximum number of pbufs on queue for loopback
  * sending for each netif (0 = disabled)
  */
-#ifndef LWIP_LOOPBACK_MAX_PBUFS
 #define LWIP_LOOPBACK_MAX_PBUFS         0
-#endif
 
 /**
  * LWIP_NETIF_LOOPBACK_MULTITHREADING: Indicates whether threading is enabled in
@@ -1149,9 +912,7 @@
  *       The packets are put on a list and netif_poll() must be called in
  *       the main application loop.
  */
-#ifndef LWIP_NETIF_LOOPBACK_MULTITHREADING
 #define LWIP_NETIF_LOOPBACK_MULTITHREADING    (!NO_SYS)
-#endif
 
 /**
  * LWIP_NETIF_TX_SINGLE_PBUF: if this is set to 1, lwIP tries to put all data
@@ -1162,9 +923,7 @@
  *
  * @todo: TCP and IP-frag do not work with this, yet:
  */
-#ifndef LWIP_NETIF_TX_SINGLE_PBUF
 #define LWIP_NETIF_TX_SINGLE_PBUF             1
-#endif /* LWIP_NETIF_TX_SINGLE_PBUF */
 
 /*
    ------------------------------------
@@ -1174,9 +933,7 @@
 /**
  * LWIP_HAVE_LOOPIF==1: Support loop interface (127.0.0.1) and loopif.c
  */
-#ifndef LWIP_HAVE_LOOPIF
 #define LWIP_HAVE_LOOPIF                0
-#endif
 
 /*
    ------------------------------------
@@ -1186,9 +943,7 @@
 /**
  * LWIP_HAVE_SLIPIF==1: Support slip interface and slipif.c
  */
-#ifndef LWIP_HAVE_SLIPIF
 #define LWIP_HAVE_SLIPIF                0
-#endif
 
 /*
    ------------------------------------
@@ -1198,147 +953,113 @@
 /**
  * TCPIP_THREAD_NAME: The name assigned to the main tcpip thread.
  */
-#ifndef TCPIP_THREAD_NAME
 #define TCPIP_THREAD_NAME              "tcpip_thread"
-#endif
 
 /**
  * TCPIP_THREAD_STACKSIZE: The stack size used by the main tcpip thread.
  * The stack size value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef TCPIP_THREAD_STACKSIZE
 #define TCPIP_THREAD_STACKSIZE          0
-#endif
 
 /**
  * TCPIP_THREAD_PRIO: The priority assigned to the main tcpip thread.
  * The priority value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef TCPIP_THREAD_PRIO
 #define TCPIP_THREAD_PRIO               1
-#endif
 
 /**
  * TCPIP_MBOX_SIZE: The mailbox size for the tcpip thread messages
  * The queue size value itself is platform-dependent, but is passed to
  * sys_mbox_new() when tcpip_init is called.
  */
-#ifndef TCPIP_MBOX_SIZE
 #define TCPIP_MBOX_SIZE                 0
-#endif
 
 /**
  * SLIPIF_THREAD_NAME: The name assigned to the slipif_loop thread.
  */
-#ifndef SLIPIF_THREAD_NAME
 #define SLIPIF_THREAD_NAME             "slipif_loop"
-#endif
 
 /**
  * SLIP_THREAD_STACKSIZE: The stack size used by the slipif_loop thread.
  * The stack size value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef SLIPIF_THREAD_STACKSIZE
 #define SLIPIF_THREAD_STACKSIZE         0
-#endif
 
 /**
  * SLIPIF_THREAD_PRIO: The priority assigned to the slipif_loop thread.
  * The priority value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef SLIPIF_THREAD_PRIO
 #define SLIPIF_THREAD_PRIO              1
-#endif
 
 /**
  * PPP_THREAD_NAME: The name assigned to the pppInputThread.
  */
-#ifndef PPP_THREAD_NAME
 #define PPP_THREAD_NAME                "pppInputThread"
-#endif
 
 /**
  * PPP_THREAD_STACKSIZE: The stack size used by the pppInputThread.
  * The stack size value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef PPP_THREAD_STACKSIZE
 #define PPP_THREAD_STACKSIZE            0
-#endif
 
 /**
  * PPP_THREAD_PRIO: The priority assigned to the pppInputThread.
  * The priority value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef PPP_THREAD_PRIO
 #define PPP_THREAD_PRIO                 1
-#endif
 
 /**
  * DEFAULT_THREAD_NAME: The name assigned to any other lwIP thread.
  */
-#ifndef DEFAULT_THREAD_NAME
 #define DEFAULT_THREAD_NAME            "lwIP"
-#endif
 
 /**
  * DEFAULT_THREAD_STACKSIZE: The stack size used by any other lwIP thread.
  * The stack size value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef DEFAULT_THREAD_STACKSIZE
 #define DEFAULT_THREAD_STACKSIZE        0
-#endif
 
 /**
  * DEFAULT_THREAD_PRIO: The priority assigned to any other lwIP thread.
  * The priority value itself is platform-dependent, but is passed to
  * sys_thread_new() when the thread is created.
  */
-#ifndef DEFAULT_THREAD_PRIO
 #define DEFAULT_THREAD_PRIO             1
-#endif
 
 /**
  * DEFAULT_RAW_RECVMBOX_SIZE: The mailbox size for the incoming packets on a
  * NETCONN_RAW. The queue size value itself is platform-dependent, but is passed
  * to sys_mbox_new() when the recvmbox is created.
  */
-#ifndef DEFAULT_RAW_RECVMBOX_SIZE
 #define DEFAULT_RAW_RECVMBOX_SIZE       0
-#endif
 
 /**
  * DEFAULT_UDP_RECVMBOX_SIZE: The mailbox size for the incoming packets on a
  * NETCONN_UDP. The queue size value itself is platform-dependent, but is passed
  * to sys_mbox_new() when the recvmbox is created.
  */
-#ifndef DEFAULT_UDP_RECVMBOX_SIZE
 #define DEFAULT_UDP_RECVMBOX_SIZE       0
-#endif
 
 /**
  * DEFAULT_TCP_RECVMBOX_SIZE: The mailbox size for the incoming packets on a
  * NETCONN_TCP. The queue size value itself is platform-dependent, but is passed
  * to sys_mbox_new() when the recvmbox is created.
  */
-#ifndef DEFAULT_TCP_RECVMBOX_SIZE
 #define DEFAULT_TCP_RECVMBOX_SIZE       0
-#endif
 
 /**
  * DEFAULT_ACCEPTMBOX_SIZE: The mailbox size for the incoming connections.
  * The queue size value itself is platform-dependent, but is passed to
  * sys_mbox_new() when the acceptmbox is created.
  */
-#ifndef DEFAULT_ACCEPTMBOX_SIZE
 #define DEFAULT_ACCEPTMBOX_SIZE         0
-#endif
 
 /*
    ----------------------------------------------
@@ -1349,31 +1070,23 @@
  * LWIP_TCPIP_CORE_LOCKING: (EXPERIMENTAL!)
  * Don't use it if you're not an active lwIP project member
  */
-#ifndef LWIP_TCPIP_CORE_LOCKING
 #define LWIP_TCPIP_CORE_LOCKING         0
-#endif
 
 /**
  * LWIP_TCPIP_CORE_LOCKING_INPUT: (EXPERIMENTAL!)
  * Don't use it if you're not an active lwIP project member
  */
-#ifndef LWIP_TCPIP_CORE_LOCKING_INPUT
 #define LWIP_TCPIP_CORE_LOCKING_INPUT   0
-#endif
 
 /**
  * LWIP_NETCONN==1: Enable Netconn API (require to use api_lib.c)
  */
-#ifndef LWIP_NETCONN
 #define LWIP_NETCONN                    0
-#endif
 
 /** LWIP_TCPIP_TIMEOUT==1: Enable tcpip_timeout/tcpip_untimeout tod create
  * timers running in tcpip_thread from another thread.
  */
-#ifndef LWIP_TCPIP_TIMEOUT
 #define LWIP_TCPIP_TIMEOUT              1
-#endif
 
 /*
    ------------------------------------
@@ -1383,72 +1096,54 @@
 /**
  * LWIP_SOCKET==1: Enable Socket API (require to use sockets.c)
  */
-#ifndef LWIP_SOCKET
 #define LWIP_SOCKET                     0
-#endif
 
 /**
  * LWIP_COMPAT_SOCKETS==1: Enable BSD-style sockets functions names.
  * (only used if you use sockets.c)
  */
-#ifndef LWIP_COMPAT_SOCKETS
 #define LWIP_COMPAT_SOCKETS             0
-#endif
 
 /**
  * LWIP_POSIX_SOCKETS_IO_NAMES==1: Enable POSIX-style sockets functions names.
  * Disable this option if you use a POSIX operating system that uses the same
  * names (read, write & close). (only used if you use sockets.c)
  */
-#ifndef LWIP_POSIX_SOCKETS_IO_NAMES
 #define LWIP_POSIX_SOCKETS_IO_NAMES     0
-#endif
 
 /**
  * LWIP_TCP_KEEPALIVE==1: Enable TCP_KEEPIDLE, TCP_KEEPINTVL and TCP_KEEPCNT
  * options processing. Note that TCP_KEEPIDLE and TCP_KEEPINTVL have to be set
  * in seconds. (does not require sockets.c, and will affect tcp.c)
  */
-#ifndef LWIP_TCP_KEEPALIVE
 #define LWIP_TCP_KEEPALIVE              1
-#endif
 
 /**
  * LWIP_SO_RCVTIMEO==1: Enable SO_RCVTIMEO processing.
  */
-#ifndef LWIP_SO_RCVTIMEO
 #define LWIP_SO_RCVTIMEO                0
-#endif
 
 /**
  * LWIP_SO_RCVBUF==1: Enable SO_RCVBUF processing.
  */
-#ifndef LWIP_SO_RCVBUF
 #define LWIP_SO_RCVBUF                  0
-#endif
 
 /**
  * If LWIP_SO_RCVBUF is used, this is the default value for recv_bufsize.
  */
-#ifndef RECV_BUFSIZE_DEFAULT
 #define RECV_BUFSIZE_DEFAULT            INT_MAX
-#endif
 
 /**
  * SO_REUSE==1: Enable SO_REUSEADDR option.
  */
-#ifndef SO_REUSE
 #define SO_REUSE                        0
-#endif
 
 /**
  * SO_REUSE_RXTOALL==1: Pass a copy of incoming broadcast/multicast packets
  * to all local matches if SO_REUSEADDR is turned on.
  * WARNING: Adds a memcpy for every packet if passing to more than one pcb!
  */
-#ifndef SO_REUSE_RXTOALL
 #define SO_REUSE_RXTOALL                0
-#endif
 
 /*
    ----------------------------------------
@@ -1458,100 +1153,75 @@
 /**
  * LWIP_STATS==1: Enable statistics collection in lwip_stats.
  */
-#ifndef LWIP_STATS
 #define LWIP_STATS                      0
-#endif
 
 #if LWIP_STATS
 
 /**
  * LWIP_STATS_DISPLAY==1: Compile in the statistics output functions.
  */
-#ifndef LWIP_STATS_DISPLAY
 #define LWIP_STATS_DISPLAY              0
-#endif
 
 /**
  * LINK_STATS==1: Enable link stats.
  */
-#ifndef LINK_STATS
 #define LINK_STATS                      1
-#endif
 
 /**
  * ETHARP_STATS==1: Enable etharp stats.
  */
-#ifndef ETHARP_STATS
 #define ETHARP_STATS                    (LWIP_ARP)
-#endif
 
 /**
  * IP_STATS==1: Enable IP stats.
  */
-#ifndef IP_STATS
 #define IP_STATS                        1
-#endif
 
 /**
  * IPFRAG_STATS==1: Enable IP fragmentation stats. Default is
  * on if using either frag or reass.
  */
-#ifndef IPFRAG_STATS
 #define IPFRAG_STATS                    (IP_REASSEMBLY || IP_FRAG)
-#endif
 
 /**
  * ICMP_STATS==1: Enable ICMP stats.
  */
-#ifndef ICMP_STATS
 #define ICMP_STATS                      1
-#endif
 
 /**
  * IGMP_STATS==1: Enable IGMP stats.
  */
-#ifndef IGMP_STATS
 #define IGMP_STATS                      (LWIP_IGMP)
-#endif
 
 /**
  * UDP_STATS==1: Enable UDP stats. Default is on if
  * UDP enabled, otherwise off.
  */
-#ifndef UDP_STATS
 #define UDP_STATS                       (LWIP_UDP)
-#endif
 
 /**
  * TCP_STATS==1: Enable TCP stats. Default is on if TCP
  * enabled, otherwise off.
  */
-#ifndef TCP_STATS
 #define TCP_STATS                       (LWIP_TCP)
-#endif
 
 /**
  * MEM_STATS==1: Enable mem.c stats.
  */
-#ifndef MEM_STATS
 #define MEM_STATS                       ((MEM_LIBC_MALLOC == 0) && (MEM_USE_POOLS == 0))
-#endif
 
 /**
  * MEMP_STATS==1: Enable memp.c pool stats.
  */
-#ifndef MEMP_STATS
 #define MEMP_STATS                      (MEMP_MEM_MALLOC == 0)
-#endif
 
 /**
  * SYS_STATS==1: Enable system stats (sem and mbox counts, etc).
  */
-#ifndef SYS_STATS
 #define SYS_STATS                       (NO_SYS == 0)
-#endif
 
-#else
+#else // !LWIP_STATS
+
 #define ETHARP_STATS                    0
 #define LINK_STATS                      0
 #define IP_STATS                        0
@@ -1575,131 +1245,87 @@
 /**
  * PPP_SUPPORT==1: Enable PPP.
  */
-#ifndef PPP_SUPPORT
 #define PPP_SUPPORT                     0
-#endif
 
 /**
  * PPPOE_SUPPORT==1: Enable PPP Over Ethernet
  */
-#ifndef PPPOE_SUPPORT
 #define PPPOE_SUPPORT                   0
-#endif
 
 /**
  * PPPOS_SUPPORT==1: Enable PPP Over Serial
  */
-#ifndef PPPOS_SUPPORT
 #define PPPOS_SUPPORT                   PPP_SUPPORT
-#endif
 
 #if PPP_SUPPORT
 
 /**
  * NUM_PPP: Max PPP sessions.
  */
-#ifndef NUM_PPP
 #define NUM_PPP                         1
-#endif
 
 /**
  * PAP_SUPPORT==1: Support PAP.
  */
-#ifndef PAP_SUPPORT
 #define PAP_SUPPORT                     0
-#endif
 
 /**
  * CHAP_SUPPORT==1: Support CHAP.
  */
-#ifndef CHAP_SUPPORT
 #define CHAP_SUPPORT                    0
-#endif
 
 /**
  * MSCHAP_SUPPORT==1: Support MSCHAP. CURRENTLY NOT SUPPORTED! DO NOT SET!
  */
-#ifndef MSCHAP_SUPPORT
 #define MSCHAP_SUPPORT                  0
-#endif
 
 /**
  * CBCP_SUPPORT==1: Support CBCP. CURRENTLY NOT SUPPORTED! DO NOT SET!
  */
-#ifndef CBCP_SUPPORT
 #define CBCP_SUPPORT                    0
-#endif
 
 /**
  * CCP_SUPPORT==1: Support CCP. CURRENTLY NOT SUPPORTED! DO NOT SET!
  */
-#ifndef CCP_SUPPORT
 #define CCP_SUPPORT                     0
-#endif
 
 /**
  * VJ_SUPPORT==1: Support VJ header compression.
  */
-#ifndef VJ_SUPPORT
 #define VJ_SUPPORT                      0
-#endif
 
 /**
  * MD5_SUPPORT==1: Support MD5 (see also CHAP).
  */
-#ifndef MD5_SUPPORT
 #define MD5_SUPPORT                     0
-#endif
 
 /*
  * Timeouts
  */
-#ifndef FSM_DEFTIMEOUT
 #define FSM_DEFTIMEOUT                  6       /* Timeout time in seconds */
-#endif
 
-#ifndef FSM_DEFMAXTERMREQS
 #define FSM_DEFMAXTERMREQS              2       /* Maximum Terminate-Request transmissions */
-#endif
 
-#ifndef FSM_DEFMAXCONFREQS
 #define FSM_DEFMAXCONFREQS              10      /* Maximum Configure-Request transmissions */
-#endif
 
-#ifndef FSM_DEFMAXNAKLOOPS
 #define FSM_DEFMAXNAKLOOPS              5       /* Maximum number of nak loops */
-#endif
 
-#ifndef UPAP_DEFTIMEOUT
 #define UPAP_DEFTIMEOUT                 6       /* Timeout (seconds) for retransmitting req */
-#endif
 
-#ifndef UPAP_DEFREQTIME
 #define UPAP_DEFREQTIME                 30      /* Time to wait for auth-req from peer */
-#endif
 
-#ifndef CHAP_DEFTIMEOUT
 #define CHAP_DEFTIMEOUT                 6       /* Timeout time in seconds */
-#endif
 
-#ifndef CHAP_DEFTRANSMITS
 #define CHAP_DEFTRANSMITS               10      /* max # times to send challenge */
-#endif
 
 /* Interval in seconds between keepalive echo requests, 0 to disable. */
-#ifndef LCP_ECHOINTERVAL
 #define LCP_ECHOINTERVAL                0
-#endif
 
 /* Number of unanswered echo requests before failure. */
-#ifndef LCP_MAXECHOFAILS
 #define LCP_MAXECHOFAILS                3
-#endif
 
 /* Max Xmit idle time (in jiffies) before resend flag char. */
-#ifndef PPP_MAXIDLEFLAG
 #define PPP_MAXIDLEFLAG                 100
-#endif
 
 /*
  * Packet sizes
@@ -1710,24 +1336,16 @@
  *    of living in lcp.h)
  */
 #define PPP_MTU                         1500     /* Default MTU (size of Info field) */
-#ifndef PPP_MAXMTU
 /* #define PPP_MAXMTU  65535 - (PPP_HDRLEN + PPP_FCSLEN) */
 #define PPP_MAXMTU                      1500 /* Largest MTU we allow */
-#endif
 #define PPP_MINMTU                      64
 #define PPP_MRU                         1500     /* default MRU = max length of info field */
 #define PPP_MAXMRU                      1500     /* Largest MRU we allow */
-#ifndef PPP_DEFMRU
 #define PPP_DEFMRU                      296             /* Try for this */
-#endif
 #define PPP_MINMRU                      128             /* No MRUs below this */
 
-#ifndef MAXNAMELEN
 #define MAXNAMELEN                      256     /* max length of hostname or name for auth */
-#endif
-#ifndef MAXSECRETLEN
 #define MAXSECRETLEN                    256     /* max length of password or secret */
-#endif
 
 #endif /* PPP_SUPPORT */
 
@@ -1739,52 +1357,38 @@
 /**
  * CHECKSUM_GEN_IP==1: Generate checksums in software for outgoing IP packets.
  */
-#ifndef CHECKSUM_GEN_IP
 #define CHECKSUM_GEN_IP                 1
-#endif
  
 /**
  * CHECKSUM_GEN_UDP==1: Generate checksums in software for outgoing UDP packets.
  */
-#ifndef CHECKSUM_GEN_UDP
 #define CHECKSUM_GEN_UDP                1
-#endif
  
 /**
  * CHECKSUM_GEN_TCP==1: Generate checksums in software for outgoing TCP packets.
  */
-#ifndef CHECKSUM_GEN_TCP
 #define CHECKSUM_GEN_TCP                1
-#endif
  
 /**
  * CHECKSUM_CHECK_IP==1: Check checksums in software for incoming IP packets.
  */
-#ifndef CHECKSUM_CHECK_IP
 #define CHECKSUM_CHECK_IP               1
-#endif
  
 /**
  * CHECKSUM_CHECK_UDP==1: Check checksums in software for incoming UDP packets.
  */
-#ifndef CHECKSUM_CHECK_UDP
 #define CHECKSUM_CHECK_UDP              1
-#endif
 
 /**
  * CHECKSUM_CHECK_TCP==1: Check checksums in software for incoming TCP packets.
  */
-#ifndef CHECKSUM_CHECK_TCP
 #define CHECKSUM_CHECK_TCP              1
-#endif
 
 /**
  * LWIP_CHECKSUM_ON_COPY==1: Calculate checksum when copying data from
  * application buffers to pbufs.
  */
-#ifndef LWIP_CHECKSUM_ON_COPY
 #define LWIP_CHECKSUM_ON_COPY           0
-#endif
 
 /*
    ---------------------------------------
@@ -1796,261 +1400,220 @@
  * compared against this value. If it is smaller, then debugging
  * messages are written.
  */
-#ifndef LWIP_DBG_MIN_LEVEL
 #define LWIP_DBG_MIN_LEVEL              LWIP_DBG_LEVEL_ALL
-#endif
 
 /**
  * LWIP_DBG_TYPES_ON: A mask that can be used to globally enable/disable
  * debug messages of certain types.
  */
-#ifndef LWIP_DBG_TYPES_ON
 #define LWIP_DBG_TYPES_ON               LWIP_DBG_OFF
-#endif
 
 /**
  * ETHARP_DEBUG: Enable debugging in etharp.c.
  */
-#ifndef ETHARP_DEBUG
 #define ETHARP_DEBUG                    LWIP_DBG_OFF
-#endif
 
 /**
  * NETIF_DEBUG: Enable debugging in netif.c.
  */
-#ifndef NETIF_DEBUG
 #define NETIF_DEBUG                     LWIP_DBG_OFF
-#endif
 
 /**
  * PBUF_DEBUG: Enable debugging in pbuf.c.
  */
-#ifndef PBUF_DEBUG
 #define PBUF_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * API_LIB_DEBUG: Enable debugging in api_lib.c.
  */
-#ifndef API_LIB_DEBUG
 #define API_LIB_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * API_MSG_DEBUG: Enable debugging in api_msg.c.
  */
-#ifndef API_MSG_DEBUG
 #define API_MSG_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * SOCKETS_DEBUG: Enable debugging in sockets.c.
  */
-#ifndef SOCKETS_DEBUG
 #define SOCKETS_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * ICMP_DEBUG: Enable debugging in icmp.c.
  */
-#ifndef ICMP_DEBUG
 #define ICMP_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * IGMP_DEBUG: Enable debugging in igmp.c.
  */
-#ifndef IGMP_DEBUG
 #define IGMP_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * INET_DEBUG: Enable debugging in inet.c.
  */
-#ifndef INET_DEBUG
 #define INET_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * IP_DEBUG: Enable debugging for IP.
  */
-#ifndef IP_DEBUG
 #define IP_DEBUG                        LWIP_DBG_OFF
-#endif
 
 /**
  * IP_REASS_DEBUG: Enable debugging in ip_frag.c for both frag & reass.
  */
-#ifndef IP_REASS_DEBUG
 #define IP_REASS_DEBUG                  LWIP_DBG_OFF
-#endif
 
 /**
  * RAW_DEBUG: Enable debugging in raw.c.
  */
-#ifndef RAW_DEBUG
 #define RAW_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * MEM_DEBUG: Enable debugging in mem.c.
  */
-#ifndef MEM_DEBUG
 #define MEM_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * MEMP_DEBUG: Enable debugging in memp.c.
  */
-#ifndef MEMP_DEBUG
 #define MEMP_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * SYS_DEBUG: Enable debugging in sys.c.
  */
-#ifndef SYS_DEBUG
 #define SYS_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * TIMERS_DEBUG: Enable debugging in timers.c.
  */
-#ifndef TIMERS_DEBUG
 #define TIMERS_DEBUG                    LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_DEBUG: Enable debugging for TCP.
  */
-#ifndef TCP_DEBUG
 #define TCP_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_INPUT_DEBUG: Enable debugging in tcp_in.c for incoming debug.
  */
-#ifndef TCP_INPUT_DEBUG
 #define TCP_INPUT_DEBUG                 LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_FR_DEBUG: Enable debugging in tcp_in.c for fast retransmit.
  */
-#ifndef TCP_FR_DEBUG
 #define TCP_FR_DEBUG                    LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_RTO_DEBUG: Enable debugging in TCP for retransmit
  * timeout.
  */
-#ifndef TCP_RTO_DEBUG
 #define TCP_RTO_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_CWND_DEBUG: Enable debugging for TCP congestion window.
  */
-#ifndef TCP_CWND_DEBUG
 #define TCP_CWND_DEBUG                  LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_WND_DEBUG: Enable debugging in tcp_in.c for window updating.
  */
-#ifndef TCP_WND_DEBUG
 #define TCP_WND_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_OUTPUT_DEBUG: Enable debugging in tcp_out.c output functions.
  */
-#ifndef TCP_OUTPUT_DEBUG
 #define TCP_OUTPUT_DEBUG                LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_RST_DEBUG: Enable debugging for TCP with the RST message.
  */
-#ifndef TCP_RST_DEBUG
 #define TCP_RST_DEBUG                   LWIP_DBG_OFF
-#endif
 
 /**
  * TCP_QLEN_DEBUG: Enable debugging for TCP queue lengths.
  */
-#ifndef TCP_QLEN_DEBUG
 #define TCP_QLEN_DEBUG                  LWIP_DBG_OFF
-#endif
 
 /**
  * UDP_DEBUG: Enable debugging in UDP.
  */
-#ifndef UDP_DEBUG
 #define UDP_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * TCPIP_DEBUG: Enable debugging in tcpip.c.
  */
-#ifndef TCPIP_DEBUG
 #define TCPIP_DEBUG                     LWIP_DBG_OFF
-#endif
 
 /**
  * PPP_DEBUG: Enable debugging for PPP.
  */
-#ifndef PPP_DEBUG
 #define PPP_DEBUG                       LWIP_DBG_OFF
-#endif
 
 /**
  * SLIP_DEBUG: Enable debugging in slipif.c.
  */
-#ifndef SLIP_DEBUG
 #define SLIP_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * DHCP_DEBUG: Enable debugging in dhcp.c.
  */
-#ifndef DHCP_DEBUG
 #define DHCP_DEBUG                      LWIP_DBG_OFF
-#endif
 
 /**
  * AUTOIP_DEBUG: Enable debugging in autoip.c.
  */
-#ifndef AUTOIP_DEBUG
 #define AUTOIP_DEBUG                    LWIP_DBG_OFF
-#endif
 
 /**
  * SNMP_MSG_DEBUG: Enable debugging for SNMP messages.
  */
-#ifndef SNMP_MSG_DEBUG
 #define SNMP_MSG_DEBUG                  LWIP_DBG_OFF
-#endif
 
 /**
  * SNMP_MIB_DEBUG: Enable debugging for SNMP MIBs.
  */
-#ifndef SNMP_MIB_DEBUG
 #define SNMP_MIB_DEBUG                  LWIP_DBG_OFF
-#endif
 
 /**
  * DNS_DEBUG: Enable debugging for DNS.
  */
-#ifndef DNS_DEBUG
 #define DNS_DEBUG                       LWIP_DBG_OFF
-#endif
 
-#ifndef LWIP_RAND
 #define LWIP_RAND()  rand()
+
+#define mem_free vPortFree
+#define mem_malloc pvPortMalloc
+#define mem_calloc pvPortCalloc
+#define mem_realloc pvPortRealloc
+#define mem_zalloc pvPortZalloc
+
+#ifndef os_malloc
+#define os_malloc(s) mem_malloc((s))
+#define os_realloc(p, s) mem_realloc((p), (s))
+#define os_zalloc(s) mem_zalloc((s))
+#define os_free(p) mem_free((p))
 #endif
 
 #define LWIP_DHCP_BOOTP_FILE 0
 
+#define IP2STR(ipaddr) ip4_addr1_16(ipaddr), \
+    ip4_addr2_16(ipaddr), \
+    ip4_addr3_16(ipaddr), \
+    ip4_addr4_16(ipaddr)
+
+#define IPSTR "%d.%d.%d.%d"
+
+struct ip_info;
+#if 0
+// this should go to antares user_interface.h
+struct ip_info {
+    struct ip_addr ip;
+    struct ip_addr netmask;
+    struct ip_addr gw;
+};
+#endif
+
+#include "user_interface.h"
+
 #endif /* __LWIP_OPT_H__ */
+
